backend/departments/views.py
from django.shortcuts import render
from rest_framework import generics
from rest_framework.response import Response
from rest_framework.decorators import api_view

# Create your views here.

# Placeholder views - will be implemented fully later
class DepartmentListView(generics.ListAPIView):
    def get(self, request):
        return Response({'message': 'Department list endpoint'})

# Add other placeholder views as needed
DepartmentCreateView = DepartmentListView
DepartmentDetailView = DepartmentListView
DepartmentUpdateView = DepartmentListView
DepartmentDeleteView = DepartmentListView
PositionListView = DepartmentListView
PositionCreateView = DepartmentListView
PositionDetailView = DepartmentListView
PositionUpdateView = DepartmentListView
PositionDeleteView = DepartmentListView
EmployeeAssignmentListView = DepartmentListView
EmployeeAssignmentCreateView = DepartmentListView
EmployeeAssignmentDetailView = DepartmentListView
EmployeeAssignmentUpdateView = DepartmentListView
DepartmentBudgetListView = DepartmentListView
DepartmentBudgetCreateView = DepartmentListView
DepartmentBudgetDetailView = DepartmentListView
DepartmentBudgetUpdateView = DepartmentListView

@api_view(['GET'])
def department_tree(request):
    return Response({'message': 'Department tree endpoint'})

@api_view(['GET'])
def department_stats(request):
    return Response({'message': 'Department stats endpoint'})

@api_view(['GET'])
def department_employees(request, department_id):
    return Response({'message': 'Department employees endpoint'})

@api_view(['POST'])
def end_assignment(request, pk):
    return Response({'message': 'End assignment endpoint'})

@api_view(['POST'])
def approve_budget(request, pk):
    return Response({'message': 'Approve budget endpoint'})

@api_view(['GET'])
def organization_chart(request):
    return Response({'message': 'Organization chart endpoint'})

@api_view(['GET'])
def performance_report(request):
    return Response({'message': 'Performance report endpoint'})

@api_view(['GET'])
def budget_utilization_report(request):
    return Response({'message': 'Budget utilization report endpoint'})
