backend/requirements.txt
asgiref==3.8.1
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
whitenoise==6.5.0

# Vector Database and AI Dependencies
qdrant-client==1.7.3
openai==1.12.0
sentence-transformers==2.7.0
transformers==4.40.0
torch==2.1.2
numpy==1.24.4
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
redis==5.0.1
celery==5.3.4

# Text processing
nltk==3.8.1
beautifulsoup4==4.12.2
